requirements.txt
# requirements.txt
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-api-python-client==2.100.0
requests==2.31.0
pyyaml==6.0.1
python-dotenv==1.0.0
pytest==7.4.2
pytest-asyncio==0.21.1
httpx==0.25.0
colorama==0.4.6
PyJWT==2.8.0
